Rework theme system: menu access, button colours, ParchmentLight rename
- Rename AppTheme.Parchment -> ParchmentLight (dark/light encoded in name)
- Rename ParchmentTheme.axaml -> ParchmentLightTheme.axaml
- Each theme now overrides SystemAccentColor (and Light1-3/Dark1-3) to replace
  the system accent colour (purple) with the theme's green/brown/blue
- Add button colour overrides (Background, PointerOver, Pressed, Foreground,
  BorderBrush) to all three theme resource dictionaries
- Move OpenSettingsCommand and ShowSettingsDialog from StartViewModel to
  MainWindowViewModel so settings are accessible from any view
- Add Menu bar to MainWindow with a Settings... item bound to the command
- Wire ShowSettingsDialog interaction in MainWindow code-behind
- Remove gear button and settings handling from StartView/StartViewModel
