+++ DMap/Models/AppTheme.cs
@@ -3,6 +3,6 @@ namespace DMap.Models;
 public enum AppTheme
 {
     Forest,
-    Parchment,
+    ParchmentLight,
     Abyss,
 }
