+++ DMap/Themes/ForestTheme.axaml
@@ -1,10 +1,29 @@
 <ResourceDictionary xmlns="https://github.com/avaloniaui"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
 
-    <!--
-        Forest: dark green theme.
-        FluentTheme is already in Dark variant — these keys add app-specific colour overrides.
-    -->
+    <!-- Forest: dark green theme (Dark variant) -->
+
+    <!-- Override FluentTheme accent colours (replaces system accent / purple) -->
+    <Color x:Key="SystemAccentColor">#4A7C59</Color>
+    <Color x:Key="SystemAccentColorLight1">#5A8C69</Color>
+    <Color x:Key="SystemAccentColorLight2">#6A9C79</Color>
+    <Color x:Key="SystemAccentColorLight3">#7AAC89</Color>
+    <Color x:Key="SystemAccentColorDark1">#3A6C49</Color>
+    <Color x:Key="SystemAccentColorDark2">#2A5C39</Color>
+    <Color x:Key="SystemAccentColorDark3">#1A4C29</Color>
+
+    <!-- Button colours -->
+    <SolidColorBrush x:Key="ButtonBackground" Color="#253324" />
+    <SolidColorBrush x:Key="ButtonBackgroundPointerOver" Color="#2E3F2D" />
+    <SolidColorBrush x:Key="ButtonBackgroundPressed" Color="#1C2719" />
+    <SolidColorBrush x:Key="ButtonForeground" Color="#E8F5E8" />
+    <SolidColorBrush x:Key="ButtonForegroundPointerOver" Color="#F0FFF0" />
+    <SolidColorBrush x:Key="ButtonForegroundPressed" Color="#E0EDE0" />
+    <SolidColorBrush x:Key="ButtonBorderBrush" Color="#3D5C3B" />
+    <SolidColorBrush x:Key="ButtonBorderBrushPointerOver" Color="#4A7C59" />
+    <SolidColorBrush x:Key="ButtonBorderBrushPressed" Color="#2E4D38" />
+
+    <!-- App-specific keys for use with {DynamicResource} in custom controls -->
     <Color x:Key="ThemeAccentColor">#4A7C59</Color>
     <SolidColorBrush x:Key="ThemeAccentBrush" Color="{StaticResource ThemeAccentColor}" />
     <SolidColorBrush x:Key="ThemeAccentSubtleBrush" Color="#2E4D38" />
