--- DMap/Themes/ParchmentTheme.axaml
@@ -1,13 +1,0 @@
-<ResourceDictionary xmlns="https://github.com/avaloniaui"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-
-    <!--
-        Parchment: warm light theme inspired by aged D&D maps.
-        FluentTheme is in Light variant — these keys add app-specific colour overrides.
-    -->
-    <Color x:Key="ThemeAccentColor">#8B4513</Color>
-    <SolidColorBrush x:Key="ThemeAccentBrush" Color="{StaticResource ThemeAccentColor}" />
-    <SolidColorBrush x:Key="ThemeAccentSubtleBrush" Color="#C68B5A" />
-    <SolidColorBrush x:Key="ThemeSurfaceBrush" Color="#EDD9A3" />
-
-</ResourceDictionary>
