+++ DMap/Views/MainWindow.axaml.cs
@@ -1,5 +1,8 @@
+using System.Reactive;
+
 using DMap.ViewModels;
 
+using ReactiveUI;
 using ReactiveUI.Avalonia;
 
 namespace DMap.Views;
@@ -9,5 +12,18 @@ public partial class MainWindow : ReactiveWindow<MainWindowViewModel>
     public MainWindow()
     {
         InitializeComponent();
+
+        this.WhenActivated(disposables =>
+        {
+            if (ViewModel is null)
+                return;
+
+            disposables(ViewModel.ShowSettingsDialog.RegisterHandler(async ctx =>
+            {
+                var dialog = new SettingsDialog { DataContext = ctx.Input };
+                await dialog.ShowDialog(this);
+                ctx.SetOutput(Unit.Default);
+            }));
+        });
     }
 }
