+++ DMap/Views/StartView.axaml
@@ -7,20 +7,13 @@
              x:Class="DMap.Views.StartView"
              x:DataType="vm:StartViewModel">
 
-    <Grid>
-        <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center" Spacing="20">
-            <TextBlock Text="DMap" FontSize="48" FontWeight="Bold" HorizontalAlignment="Center" />
-            <TextBlock Text="D&amp;D Map Tool" FontSize="16" HorizontalAlignment="Center" Opacity="0.6" />
-            <Button Content="Dungeon Master" Command="{Binding StartAsDmCommand}"
-                    Width="200" Height="60" HorizontalContentAlignment="Center" />
-            <Button Content="Player" Command="{Binding StartAsPlayerCommand}"
-                    Width="200" Height="60" HorizontalContentAlignment="Center" />
-        </StackPanel>
-
-        <Button Content="⚙" Command="{Binding OpenSettingsCommand}"
-                HorizontalAlignment="Right" VerticalAlignment="Top"
-                Margin="0,8,8,0" Width="36" Height="36"
-                ToolTip.Tip="Settings" />
-    </Grid>
+    <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center" Spacing="20">
+        <TextBlock Text="DMap" FontSize="48" FontWeight="Bold" HorizontalAlignment="Center" />
+        <TextBlock Text="D&amp;D Map Tool" FontSize="16" HorizontalAlignment="Center" Opacity="0.6" />
+        <Button Content="Dungeon Master" Command="{Binding StartAsDmCommand}"
+                Width="200" Height="60" HorizontalContentAlignment="Center" />
+        <Button Content="Player" Command="{Binding StartAsPlayerCommand}"
+                Width="200" Height="60" HorizontalContentAlignment="Center" />
+    </StackPanel>
 
 </UserControl>
